Bulk Upload Because I Forgot To Do It Regularly
Database Handler has been updated so that the handler function should be able to manage all data that is passed into it with the user specifying what they want to do. E.g., Edit, remove, add, view, etc.

Also facial detection has been updated so that the camera is now what is displayed, along with a square that hovers around the face of a person that is looking at the camera.

Lastly, gui hasn't really had much of an update yet, but this is to come in due time, after all unittesting has been designed for the program.