Security and Stability Update
Split up authentication to another handler, allowing the implementation of safer authentication.

Moreover, more exception handling and Custom Exceptions to help prevent runtime crashes as well as making the errors that are occurring more detailed.

Lastly, patch for when an employee is edited. When an employee has their name edited, ensuring that the encoded.bin files are always findable from the database data.